Add the premice of header insersion for itp files
Add a non working function which aims to insert a comment line
describing the format of the current section in a top/itp file.

The function is not working because the dictionnary define multiple
times the same key. This is due to the fact that a section can actually
have several different formats and I still need a way to discriminate
between them.